SplitScreen Refactored
